feat(client):do not show resource selector and breadcrumbs on platform console
