+++ packages/amplication-client/src/Workspaces/WorkspaceHeader/WorkspaceNavigation.scss
@@ -1,0 +1,7 @@
+@import "../../style/index.scss";
+
+.workspace-navigation {
+  &__separator {
+    color: var(--gray-20);
+  }
+}
